Fix tests for postgres 11
Postgres 11 closes the connection after certain errors
(syntax errors). In this case, the cleanup code
of query_fixtures cannot run.
Note that not cleaning up explicitly via DROP is not really
an issue (of accumulating test tables), because the tests
do not issue a COMMIT, so the tables are never persisted.